+++ R/wrap_add_prior_information.R
@@ -34,7 +34,6 @@
 #'
 #' @export
 #'
-#' @importFrom testthat expect_true
 #' @importFrom purrr discard list_modify
 add_prior_information <- function(
   dataset,
@@ -67,45 +66,45 @@ add_prior_information <- function(
   ) %>% discard(is.null)
 
   if (!is.null(start_id)) {
-    testthat::expect_true(all(start_id %in% dataset$cell_ids))
+    assert_that(all(start_id %in% dataset$cell_ids))
   }
   if (!is.null(end_id)) {
-    testthat::expect_true(all(start_id %in% dataset$cell_ids))
+    assert_that(all(start_id %in% dataset$cell_ids))
   }
   if (!is.null(groups_id)) {
     if(is.vector(groups_id)) {
-      testthat::expect_true(!is.null(names(groups_id)))
+      assert_that(!is.null(names(groups_id)))
       groups_id <- enframe(groups_id, "cell_id", "group_id")
       prior_information$groups_id <- groups_id
     }
-    testthat::expect_true(is.data.frame(groups_id))
-    testthat::expect_setequal(colnames(groups_id), c("cell_id", "group_id"))
-    testthat::expect_setequal(groups_id$cell_id, dataset$cell_ids)
+    assert_that(is.data.frame(groups_id))
+    assert_that(setequal(colnames(groups_id), c("cell_id", "group_id")))
+    assert_that(setequal(groups_id$cell_id, dataset$cell_ids))
   }
   if (!is.null(groups_network)) {
-    testthat::expect_true(!is.null(groups_id))
-    testthat::expect_setequal(colnames(groups_network), c("from", "to"))
-    testthat::expect_true(all(groups_id$group_id %in% c(groups_network$to, groups_network$from)))
+    assert_that(!is.null(groups_id))
+    assert_that(setequal(colnames(groups_network), c("from", "to")))
+    assert_that(all(groups_id$group_id %in% c(groups_network$to, groups_network$from)))
   }
   if (!is.null(features_id)) {
-    testthat::expect_true(is_wrapper_with_expression(dataset))
-    testthat::expect_true(all(features_id %in% colnames(dataset$counts)))
+    assert_that(is_wrapper_with_expression(dataset))
+    assert_that(all(features_id %in% colnames(dataset$counts)))
   }
   if (!is.null(timecourse_continuous)) {
-    testthat::expect_true(all(is.numeric(timecourse_continuous)))
-    testthat::expect_setequal(dataset$cell_ids, names(timecourse_continuous))
+    assert_that(all(is.numeric(timecourse_continuous)))
+    assert_that(setequal(dataset$cell_ids, names(timecourse_continuous)))
 
     prior_information$timecourse_continuous <- timecourse_continuous[dataset$cell_ids]
   }
   if (!is.null(timecourse_discrete)) {
-    testthat::expect_true(is.numeric(timecourse_discrete))
-    testthat::expect_setequal(dataset$cell_ids, names(timecourse_discrete))
+    assert_that(is.numeric(timecourse_discrete))
+    assert_that(setequal(dataset$cell_ids, names(timecourse_discrete)))
 
     prior_information$timecourse_discrete <- timecourse_discrete[dataset$cell_ids]
   }
   if (!is.null(dimred)) {
-    testthat::expect_true(is.matrix(dimred))
-    testthat::expect_setequal(dataset$cell_ids, rownames(dimred))
+    assert_that(is.matrix(dimred))
+    assert_that(setequal(dataset$cell_ids, rownames(dimred)))
 
     prior_information$dimred <- dimred[dataset$cell_ids, ]
   }